+++ swiftwinrt/file_writers.h
@@ -237,6 +237,10 @@ namespace swiftwinrt
 
         w.write("%", w.filter.bind_each<write_struct_bridgeable>(members.structs));
 
+        // MakeFromAbi has to be in impl file (or main file) otherwise they get stripped away
+        w.write("%", w.filter.bind_each<write_make_from_abi>(members.interfaces));
+        w.write("%", w.filter.bind_each<write_make_from_abi>(members.classes));
+
         w.swap();
         write_preamble(w, /* swift_code: */ true);
 
@@ -273,22 +277,5 @@ namespace swiftwinrt
         w.swap();
         write_preamble(w, /* swift_code: */ true);
         w.save_file("Generics");
-    }
-
-    static void write_module_make_from_abi(std::string_view const& module, type_cache const& members, include_only_used_filter const& filter)
-    {
-        writer w;
-        w.filter = filter;
-        w.support = settings.support;
-        w.c_mod = settings.get_c_module_name();
-        w.type_namespace = module;
-        w.swift_module = module;
-        w.cache = members.cache;
-        w.write("%", w.filter.bind_each<write_make_from_abi>(members.interfaces));
-        w.write("%", w.filter.bind_each<write_make_from_abi>(members.classes));
-
-        w.swap();
-        write_preamble(w, /* swift_code: */ true);
-        w.save_file("MakeFromAbi");
     }
 }
