+++ tests/test_component/Sources/test_component/Windows.Storage.FileProperties+Impl.swift
@@ -109,3 +109,60 @@ public enum __IMPL_Windows_Storage_FileProperties {
     }
 
 }
+@_spi(WinRTInternal)
+public class IStorageItemExtraPropertiesMaker: MakeFromAbi {
+    public typealias SwiftType = AnyIStorageItemExtraProperties
+    public static func from(abi: test_component.IInspectable) -> SwiftType {
+        let swiftAbi: __ABI_Windows_Storage_FileProperties.IStorageItemExtraProperties = try! abi.QueryInterface()
+        return __IMPL_Windows_Storage_FileProperties.IStorageItemExtraPropertiesBridge.from(abi: RawPointer(swiftAbi))!
+    }
+}
+@_spi(WinRTInternal)
+public class BasicPropertiesMaker: MakeFromAbi {
+    public typealias SwiftType = BasicProperties
+    public static func from(abi: test_component.IInspectable) -> SwiftType {
+        return BasicProperties(fromAbi: abi)
+    }
+}
+@_spi(WinRTInternal)
+public class DocumentPropertiesMaker: MakeFromAbi {
+    public typealias SwiftType = DocumentProperties
+    public static func from(abi: test_component.IInspectable) -> SwiftType {
+        return DocumentProperties(fromAbi: abi)
+    }
+}
+@_spi(WinRTInternal)
+public class ImagePropertiesMaker: MakeFromAbi {
+    public typealias SwiftType = ImageProperties
+    public static func from(abi: test_component.IInspectable) -> SwiftType {
+        return ImageProperties(fromAbi: abi)
+    }
+}
+@_spi(WinRTInternal)
+public class MusicPropertiesMaker: MakeFromAbi {
+    public typealias SwiftType = MusicProperties
+    public static func from(abi: test_component.IInspectable) -> SwiftType {
+        return MusicProperties(fromAbi: abi)
+    }
+}
+@_spi(WinRTInternal)
+public class StorageItemContentPropertiesMaker: MakeFromAbi {
+    public typealias SwiftType = StorageItemContentProperties
+    public static func from(abi: test_component.IInspectable) -> SwiftType {
+        return StorageItemContentProperties(fromAbi: abi)
+    }
+}
+@_spi(WinRTInternal)
+public class StorageItemThumbnailMaker: MakeFromAbi {
+    public typealias SwiftType = StorageItemThumbnail
+    public static func from(abi: test_component.IInspectable) -> SwiftType {
+        return StorageItemThumbnail(fromAbi: abi)
+    }
+}
+@_spi(WinRTInternal)
+public class VideoPropertiesMaker: MakeFromAbi {
+    public typealias SwiftType = VideoProperties
+    public static func from(abi: test_component.IInspectable) -> SwiftType {
+        return VideoProperties(fromAbi: abi)
+    }
+}
